[RF] Refactor BatchMode likelihood creation code
When creating a likelihood for the batch mode, there are two distinct
steps:

  1. Create the computation graph for the likelihood
  2. Instantiate a RooFitDriver instance to evaluate the likelihood

The first step will the shared with the new RooFit Automatic
Differentiation (AD) backend. To make this easy, the likelihood creation
for the BatchMode is refactored such that both steps listed above don't
happen in the same function anymore.

After this PR, it will only require small changes to make the codegen+AD
backend available to `RooAbsPdf::createNLL()`.
