+++ roofit/roofitcore/src/BatchModeHelpers.cxx
@@ -19,14 +19,12 @@
 #include <RooCategory.h>
 #include <RooConstraintSum.h>
 #include <RooDataSet.h>
-#include "RooFitDriver.h"
 #include "RooNLLVarNew.h"
 #include <RooRealVar.h>
 #include <RooSimultaneous.h>
 
 #include <string>
 
-using ROOT::Experimental::RooFitDriver;
 using ROOT::Experimental::RooNLLVarNew;
 
 namespace {
@@ -74,96 +72,12 @@ std::unique_ptr<RooAbsArg> createSimultaneousNLL(RooSimultaneous const &simPdf, 
    return nll;
 }
 
-class RooAbsRealWrapper final : public RooAbsReal {
-public:
-   RooAbsRealWrapper(std::unique_ptr<RooFitDriver> driver, std::string const &rangeName, RooSimultaneous const *simPdf,
-                     bool takeGlobalObservablesFromData)
-      : RooAbsReal{"RooFitDriverWrapper", "RooFitDriverWrapper"}, _driver{std::move(driver)},
-        _topNode("topNode", "top node", this, _driver->topNode()), _rangeName{rangeName}, _simPdf{simPdf},
-        _takeGlobalObservablesFromData{takeGlobalObservablesFromData}
-   {
-   }
-
-   RooAbsRealWrapper(const RooAbsRealWrapper &other, const char *name = nullptr)
-      : RooAbsReal{other, name}, _driver{other._driver},
-        _topNode("topNode", this, other._topNode), _data{other._data}, _parameters{other._parameters},
-        _rangeName{other._rangeName}, _simPdf{other._simPdf}, _takeGlobalObservablesFromData{
-                                                                 other._takeGlobalObservablesFromData}
-   {
-   }
-
-   TObject *clone(const char *newname) const override { return new RooAbsRealWrapper(*this, newname); }
-
-   double defaultErrorLevel() const override { return _driver->topNode().defaultErrorLevel(); }
-
-   bool getParameters(const RooArgSet *observables, RooArgSet &outputSet, bool /*stripDisconnected*/) const override
-   {
-      outputSet.add(_parameters);
-      if (observables) {
-         outputSet.remove(*observables);
-      }
-      // If we take the global observables as data, we have to return these as
-      // parameters instead of the parameters in the model. Otherwise, the
-      // constant parameters in the fit result that are global observables will
-      // not have the right values.
-      if (_takeGlobalObservablesFromData && _data->getGlobalObservables()) {
-         outputSet.replace(*_data->getGlobalObservables());
-      }
-      return false;
-   }
-
-   bool setData(RooAbsData &data, bool /*cloneData*/) override
-   {
-      _data = &data;
-
-      // Figure out what are the parameters for the current dataset
-      _parameters.clear();
-      RooArgSet params;
-      _driver->topNode().getParameters(_data->get(), params, true);
-      for (RooAbsArg *param : params) {
-         if (!param->getAttribute("__obs__")) {
-            _parameters.add(*param);
-         }
-      }
-
-      _driver->setData(*_data, _rangeName, _simPdf, /*skipZeroWeights=*/true, _takeGlobalObservablesFromData);
-      return true;
-   }
-
-   double getValV(const RooArgSet *) const override { return evaluate(); }
-
-   void applyWeightSquared(bool flag) override
-   {
-      const_cast<RooAbsReal &>(_driver->topNode()).applyWeightSquared(flag);
-   }
-
-   void printMultiline(std::ostream &os, Int_t /*contents*/, bool /*verbose*/ = false,
-                       TString /*indent*/ = "") const override
-   {
-      _driver->print(os);
-   }
-
-protected:
-   double evaluate() const override { return _driver ? _driver->getVal() : 0.0; }
-
-private:
-   std::shared_ptr<RooFitDriver> _driver;
-   RooRealProxy _topNode;
-   RooAbsData *_data = nullptr;
-   RooArgSet _parameters;
-   std::string _rangeName;
-   RooSimultaneous const *_simPdf = nullptr;
-   const bool _takeGlobalObservablesFromData;
-};
-
 } // namespace
 
 std::unique_ptr<RooAbsReal>
-RooFit::BatchModeHelpers::createNLL(std::unique_ptr<RooAbsPdf> &&pdf, RooAbsData &data,
-                                    std::unique_ptr<RooAbsReal> &&constraints, std::string const &rangeName,
-                                    RooArgSet const &projDeps, bool isExtended, double integrateOverBinsPrecision,
-                                    RooFit::BatchModeOption batchMode, RooFit::OffsetMode offset,
-                                    bool takeGlobalObservablesFromData)
+RooFit::BatchModeHelpers::createNLL(RooAbsPdf &pdf, RooAbsData &data, std::unique_ptr<RooAbsReal> &&constraints,
+                                    std::string const &rangeName, RooArgSet const &projDeps, bool isExtended,
+                                    double integrateOverBinsPrecision, RooFit::OffsetMode offset)
 {
    if (constraints) {
       // Redirect the global observables to the ones from the dataset if applicable.
@@ -175,18 +89,18 @@ RooFit::BatchModeHelpers::createNLL(std::unique_ptr<RooAbsPdf> &&pdf, RooAbsData
    }
 
    RooArgSet observables;
-   pdf->getObservables(data.get(), observables);
+   pdf.getObservables(data.get(), observables);
    observables.remove(projDeps, true, true);
 
-   oocxcoutI(pdf.get(), Fitting) << "RooAbsPdf::fitTo(" << pdf->GetName()
-                                 << ") fixing normalization set for coefficient determination to observables in data"
-                                 << "\n";
-   pdf->fixAddCoefNormalization(observables, false);
+   oocxcoutI(&pdf, Fitting) << "RooAbsPdf::fitTo(" << pdf.GetName()
+                            << ") fixing normalization set for coefficient determination to observables in data"
+                            << "\n";
+   pdf.fixAddCoefNormalization(observables, false);
 
    // Deal with the IntegrateBins argument
    RooArgList binSamplingPdfs;
-   std::unique_ptr<RooAbsPdf> wrappedPdf = RooBinSamplingPdf::create(*pdf, data, integrateOverBinsPrecision);
-   RooAbsPdf &finalPdf = wrappedPdf ? *wrappedPdf : *pdf;
+   std::unique_ptr<RooAbsPdf> wrappedPdf = RooBinSamplingPdf::create(pdf, data, integrateOverBinsPrecision);
+   RooAbsPdf &finalPdf = wrappedPdf ? *wrappedPdf : pdf;
    if (wrappedPdf) {
       binSamplingPdfs.addOwned(std::move(wrappedPdf));
    }
@@ -206,55 +120,10 @@ RooFit::BatchModeHelpers::createNLL(std::unique_ptr<RooAbsPdf> &&pdf, RooAbsData
       nllTerms.addOwned(std::move(constraints));
    }
 
-   std::string nllName = std::string("nll_") + pdf->GetName() + "_" + data.GetName();
+   std::string nllName = std::string("nll_") + pdf.GetName() + "_" + data.GetName();
    auto nll = std::make_unique<RooAddition>(nllName.c_str(), nllName.c_str(), nllTerms);
    nll->addOwnedComponents(std::move(binSamplingPdfs));
    nll->addOwnedComponents(std::move(nllTerms));
 
-   auto driver = std::make_unique<RooFitDriver>(*nll, batchMode);
-
-   auto driverWrapper =
-      std::make_unique<RooAbsRealWrapper>(std::move(driver), rangeName, simPdf, takeGlobalObservablesFromData);
-   driverWrapper->setData(data, false);
-   driverWrapper->addOwnedComponents(std::move(nll));
-   driverWrapper->addOwnedComponents(std::move(pdf));
-
-   return driverWrapper;
-}
-
-void RooFit::BatchModeHelpers::logArchitectureInfo(RooFit::BatchModeOption batchMode)
-{
-   // We have to exit early if the message stream is not active. Otherwise it's
-   // possible that this function skips logging because it thinks it has
-   // already logged, but actually it didn't.
-   if (!RooMsgService::instance().isActive(static_cast<RooAbsArg *>(nullptr), RooFit::Fitting, RooFit::INFO)) {
-      return;
-   }
-
-   // Don't repeat logging architecture info if the batchMode option didn't change
-   {
-      // Second element of pair tracks whether this function has already been called
-      static std::pair<RooFit::BatchModeOption, bool> lastBatchMode;
-      if (lastBatchMode.second && lastBatchMode.first == batchMode)
-         return;
-      lastBatchMode = {batchMode, true};
-   }
-
-   auto log = [](std::string_view message) {
-      oocxcoutI(static_cast<RooAbsArg *>(nullptr), Fitting) << message << std::endl;
-   };
-
-   if (batchMode == RooFit::BatchModeOption::Cuda && !RooBatchCompute::dispatchCUDA) {
-      throw std::runtime_error(std::string("In: ") + __func__ + "(), " + __FILE__ + ":" + __LINE__ +
-                               ": Cuda implementation of the computing library is not available\n");
-   }
-   if (RooBatchCompute::dispatchCPU->architecture() == RooBatchCompute::Architecture::GENERIC) {
-      log("using generic CPU library compiled with no vectorizations");
-   } else {
-      log(std::string("using CPU computation library compiled with -m") +
-          RooBatchCompute::dispatchCPU->architectureName());
-   }
-   if (batchMode == RooFit::BatchModeOption::Cuda) {
-      log("using CUDA computation library");
-   }
+   return nll;
 }
