+++ roofit/roofitcore/src/RooFitDriver.h
@@ -14,9 +14,11 @@
 #ifndef RooFit_RooFitDriver_h
 #define RooFit_RooFitDriver_h
 
+#include <RooAbsData.h>
 #include "RooFit/Detail/DataMap.h"
 #include <RooGlobalFunc.h>
 #include <RooHelpers.h>
+#include <RooRealProxy.h>
 #include "RooFit/Detail/Buffers.h"
 
 #include <chrono>
@@ -24,9 +26,7 @@
 #include <stack>
 
 class RooAbsArg;
-class RooAbsReal;
 class RooAbsCategory;
-class RooAbsData;
 class RooSimultaneous;
 
 namespace ROOT {
@@ -89,6 +89,47 @@ private:
 
    // RAII structures to reset state of computation graph after driver destruction
    std::stack<RooHelpers::ChangeOperModeRAII> _changeOperModeRAIIs;
+};
+
+class RooAbsRealWrapper final : public RooAbsReal {
+public:
+   RooAbsRealWrapper(std::unique_ptr<RooFitDriver> driver, std::string const &rangeName, RooSimultaneous const *simPdf,
+                     bool takeGlobalObservablesFromData);
+
+   RooAbsRealWrapper(const RooAbsRealWrapper &other, const char *name = nullptr);
+
+   TObject *clone(const char *newname) const override { return new RooAbsRealWrapper(*this, newname); }
+
+   double defaultErrorLevel() const override { return _driver->topNode().defaultErrorLevel(); }
+
+   bool getParameters(const RooArgSet *observables, RooArgSet &outputSet, bool stripDisconnected) const override;
+
+   bool setData(RooAbsData &data, bool cloneData) override;
+
+   double getValV(const RooArgSet *) const override { return evaluate(); }
+
+   void applyWeightSquared(bool flag) override
+   {
+      const_cast<RooAbsReal &>(_driver->topNode()).applyWeightSquared(flag);
+   }
+
+   void printMultiline(std::ostream &os, Int_t /*contents*/, bool /*verbose*/ = false,
+                       TString /*indent*/ = "") const override
+   {
+      _driver->print(os);
+   }
+
+protected:
+   double evaluate() const override { return _driver ? _driver->getVal() : 0.0; }
+
+private:
+   std::shared_ptr<RooFitDriver> _driver;
+   RooRealProxy _topNode;
+   RooAbsData *_data = nullptr;
+   RooArgSet _parameters;
+   std::string _rangeName;
+   RooSimultaneous const *_simPdf = nullptr;
+   const bool _takeGlobalObservablesFromData;
 };
 
 } // end namespace Experimental
